requirements.txt
discord.py==2.3.2
Flask==3.1.1
python-dotenv==1.1.1
Requests==2.32.4
aiohttp==3.12.14
ruff==0.12.9
black==25.1.0
asyncio==4.0.0
audioop-lts==0.2.2